refactor actions in welcome
