+++ frontend/src/gql/gql_mutation.js
@@ -11,6 +11,7 @@ export const CreateRoomMutation = gql`
   }
 `;
 
+
 export const AddPlayerMutation = gql`
   mutation($code: String!, $username: String!) {
     addPlayer(code: $code, username: $username) {
